Check if the ~/.phpctl directory exists before removing it
There is a possibility that the user already has a `.phpctl` directory in the home folder, and although the risk is low, it is not zero.

This pull request changes the behavior to check if the folder exists. If it does, then it asks the user for confirmation before deleting it.